requirements.txt
fastapi
uvicorn
pandas
requests
seaborn
matplotlib
numpy
tensorflow-cpu
image
dash
plotly.express

